update hook to take into account case where key/credentils.js is not available

+++ hooks/copyCredentials.js
@@ -1,3 +1,7 @@
+/* Copies JS file keys/credentials.js to www/js/ directory.
+This is used because the dir keys/ is not git tracked as it contains sensible information */
+
+const fs = require('fs')
 const fse = require('fs-extra')
 const path = require('path')
 
@@ -11,11 +15,19 @@ module.exports = function (context) {
   var fileOriginFullPath = path.join(projectRoot, 'keys', 'credentials.js')
   var fileDestFullPath = path.join(projectRoot, 'www', 'js', 'credentials.js')
 
-  fse.copySync(fileOriginFullPath, fileDestFullPath)
+  try {
+    if (fs.existsSync(fileOriginFullPath)) { // file exists
+      fse.copySync(fileOriginFullPath, fileDestFullPath)
 
-  const consoleMsg = 'copied ' +
-    path.relative(projectRoot, fileOriginFullPath) + ' -> ' +
-    path.relative(projectRoot, fileDestFullPath)
+      const consoleMsg = 'copied ' +
+        path.relative(projectRoot, fileOriginFullPath) + ' -> ' +
+        path.relative(projectRoot, fileDestFullPath)
 
-  console.log(twoSpaces + consoleMsg)
+      console.log(twoSpaces + consoleMsg)
+    } else { // file does no exist
+      console.log(`${twoSpaces}File ${path.relative(context.opts.projectRoot, fileOriginFullPath)} does not exist, skipping...`)
+    }
+  } catch (err) {
+    console.error(err)
+  }
 }
